fix(header): collapse header + NavBar into single 60px row
Merges the desktop nav strip into the global header, splits NavBar into
HeaderNavLinks (desktop, inline in header) and MobileNavBar (mobile bottom
tabs only). Adds a conditional Bookmarks tab for signed-in users and drops
the now-duplicate Bookmarks entry from AccountMenu.

Part of #74.
