+++ app/canvas/CanvasPageClient.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect, useRef, Suspense } from "react";
 import { useSearchParams } from "next/navigation";
 import dynamic from "next/dynamic";
 import { Header } from "@/components/layout/Header";
-import { NavBar } from "@/components/layout/NavBar";
+import { MobileNavBar } from "@/components/layout/MobileNavBar";
 import { SearchDialog } from "@/components/search/SearchDialog";
 import { EmptyState } from "@/components/canvas/EmptyState";
 import { CanvasTour } from "@/components/canvas/CanvasTour";
@@ -101,7 +101,7 @@ export function CanvasPageClient() {
       </Suspense>
 
       <Header onSearchOpen={() => setSearchOpen(true)} />
-      <NavBar />
+      <MobileNavBar />
 
       <main className="flex-1 relative overflow-hidden">
         {nodes.length === 0 && <EmptyState onSearchOpen={() => setSearchOpen(true)} />}
