+++ app/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { redirect } from "next/navigation";
 import { LandingHeader } from "@/components/layout/LandingHeader";
-import { NavBar } from "@/components/layout/NavBar";
+import { MobileNavBar } from "@/components/layout/MobileNavBar";
 import { HomeView } from "@/components/home/HomeView";
 import { getVerseOfDay } from "@/lib/quran/verse-of-day";
 
@@ -32,7 +32,7 @@ export default async function Home({
   return (
     <div className="flex h-dvh flex-col overflow-hidden bg-bg">
       <LandingHeader />
-      <NavBar />
+      <MobileNavBar />
 
       <HomeView verse={verse} />
     </div>
